rename: 쿠폰 발급 관련 컴포넌트 IssuedCoupon 폴더로 이동
