+++ src/pages/IssuedCouponPage.tsx
@@ -1,5 +1,5 @@
-import IssuedCouponHeader from "@/components/common/Header/IssuedCouponHeader";
-import IssuedCouponInfoTable from "@/components/common/Table/IssuedCouponInfoTable";
+import IssuedCouponHeader from "@/components/IssuedCoupon/IssuedCouponHeader";
+import IssuedCouponInfoTable from "@/components/IssuedCoupon/IssuedCouponInfoTable";
 import Title from "@/components/common/Title";
 import IssuedCouponContextProvider from "@/contexts/IssuedCouponContext";
